\name{cpkHistogram}
\alias{cpkHistogram}
\title{cpkHistogram}
\usage{
  cpkHistogram(myData , LoLimit, HiLimit, sTitle)
}
\arguments{
  \item{myData}{Dataset}

  \item{LoLimit}{Lo Limit}

  \item{HiLimit}{Hi Limit}
  
  \item{sTitle}{title for the chart}
}
\value{ggplot object.}
\description{
  Creates an area plot of  data.
}

